Struct spartan_lib::core::db::TreeDatabase

pub struct TreeDatabase<M> where
    M: Identifiable + Sortable,
    <M as Identifiable>::Id: Hash,  { /* fields omitted */ }

Tree-based database

Used by default server implementation.

Internally, contains message storage, and index for fast lookups, thus improving performance in comparison with VecDatabase

TreeDatabase heavily relies on correct M implementation of Sortable as only first element of index is used to check if there are any available messages in queue.